Rename 'glue' module to 'aws-glue'
