+++ aws-glue/src/it/scala/dev/chronicles/awsglue/GlueMetastoreIntegrationSpec.scala
@@ -1,4 +1,4 @@
-package dev.chronicles.glue
+package dev.chronicles.awsglue
 
 import java.net.URI
